squash! Fix warning for deprecated pytest feature
This now uses iter_markers, which is necessary when parametrize
is used twice. There is a legitimate use case for this: when you
want to test all combinations of multiple parametrizations
rather than testing a set number of values for some variables.

Also added a test.

To implement the test, I made a small change to the Plotter class,
as the filename of the plot was not originally available to
the test function. This change makes it available with no other
changes to functionality, as far as I can tell.