Deactivate progress bar in cache benchmark.
Using a progress bar starts a thread that sleeps for 0.1s for every
update. Joining the thread after completing the build thus takes up
to 0.1s too which swamps any build performance improvements for the
given test cases. By deactivating the progress bar we get the raw
build performance. Outside the tests added 0.1s are not noticeable
to the user and thus do not need to be changed.

Fixes #1447.